RadioOption prop types passed from RadioGroup are now declared required.
Separated the RadioOption used in consumer code from the actual one which declares all the required props. The former closes over functions defined by RadioGroup and renders the latter internally.

Used function as a child pattern to solve this because Ken.
Admittedly, made a bit of mess along way though.